Fixed merging schemas with different entries
